Stable extraction of bottomSheet widget to global context
